% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{all_strains}
\alias{all_strains}
\title{Hemp strains with cnoid profile and effects profile.}
\format{
An object of class \code{list} of length 12.
}
\usage{
all_strains
}
\description{
Hemp strains with cnoid profile and effects profile.
}
\details{
The cnoid profile should be
percentages ranging from 0 to 100. The data
came from https://flowercompany.com/, and the
strain effect data came from scraped data from
the internal https://leafly.com/ API.

Additional tools for scraping:
\itemize{
\item https://github.com/tytydraco/ll
Graphical tool to analyze internal Leafly API
data.
\item https://github.com/tytydraco/leafly-strain-scraper/
Command line tool to scrape the Leafly API
into a JSON document.
}
}
\keyword{datasets}
