% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim.r
\name{anim_cnoids}
\alias{anim_cnoids}
\title{Generate a radar animation of the cnoids
between two effect profiles.}
\usage{
anim_cnoids(frames, start_values, end_values)
}
\arguments{
\item{frames}{The number of frames to animate (60fps).}

\item{start_values}{The starting effect values.}

\item{end_values}{The ending effect values.}
}
\description{
The values should sum to 100.
}
\details{
We normalize the values here because otherwise
it is too challenging to see the impact of
changes. However, we lose the intensity of the
cnoid profile effects as a consequence.
}
\examples{
anim_cnoids(
  120,
  c(100, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
  c(50, 0, 0, 50, 0, 0, 0, 0, 0, 0, 0, 0, 0)
)

}
