% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{build_all_strains_matrix}
\alias{build_all_strains_matrix}
\title{Produce a matrix from the data in \link{all_strains}.}
\usage{
build_all_strains_matrix()
}
\value{
The matrix of all strains and effects.
}
\description{
Produce a matrix from the data in \link{all_strains}.
}
