% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{build_cnoid_effect_correlation}
\alias{build_cnoid_effect_correlation}
\title{Compute the correlation matrix for all strains.}
\usage{
build_cnoid_effect_correlation()
}
\value{
The matrix for the correlations between cnoids
and effects.
}
\description{
Compute the correlation matrix for all strains.
}
