% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.r
\docType{data}
\name{categories_cnoids}
\alias{categories_cnoids}
\title{The list of cnoids.}
\format{
An object of class \code{character} of length 6.
}
\usage{
categories_cnoids
}
\description{
The list of cnoids.
}
\keyword{datasets}
