% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.r
\docType{data}
\name{categories_effects}
\alias{categories_effects}
\title{The list of effects.}
\format{
An object of class \code{character} of length 13.
}
\usage{
categories_effects
}
\description{
The list of effects.
}
\keyword{datasets}
