% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_gen.r
\name{get_cached_correl_matrix}
\alias{get_cached_correl_matrix}
\title{Gives the cached correlation matrix.}
\usage{
get_cached_correl_matrix()
}
\value{
The cached correlation matrix.
}
\description{
Gives the cached correlation matrix.
}
