% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{graph_cnoids}
\alias{graph_cnoids}
\title{Generate a radar of the cnoids profile.}
\usage{
graph_cnoids(values)
}
\arguments{
\item{values}{The starting cnoid values.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         The graph.
}\if{html}{\out{</div>}}
}
\description{
The values should sum to 100.
}
\details{
We normalize the values here because otherwise
it is too challenging to see the impact of
changes. However, we lose the intensity of the
cnoid profile effects as a consequence.
}
\examples{
graph_cnoids(
  c(100, 0, 0, 0, 0, 0)
)

}
