% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{graph_effects}
\alias{graph_effects}
\title{Generate a radar of the effects profile.}
\usage{
graph_effects(values)
}
\arguments{
\item{values}{The starting cnoid values.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         The graph.
}\if{html}{\out{</div>}}
}
\description{
The values should sum to 100.
}
\examples{
graph_effects(
  c(100, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
)

}
