% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{lerp}
\alias{lerp}
\title{Simple linear interpolation between two values.}
\usage{
lerp(start, end, fraction)
}
\arguments{
\item{start}{The initial value.}

\item{end}{The final value.}

\item{fraction}{A fractional value between 0 and 1.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The interpolated value.
}\if{html}{\out{</div>}}
}
\description{
Simple linear interpolation between two values.
}
\examples{
lerp(0, 100, 0.5)

}
