% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{make_cached_correl_matrix}
\alias{make_cached_correl_matrix}
\title{Helps cache the building of the correlation
matrix.}
\usage{
make_cached_correl_matrix()
}
\value{
A function that builds or returns the
correlation matrix.
}
\description{
Helps cache the building of the correlation
matrix.
}
