% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{n_effects}
\alias{n_effects}
\title{The total number of effects in \link{all_strains}.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
n_effects
}
\description{
The total number of effects in \link{all_strains}.
}
\keyword{datasets}
