% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{normalize}
\alias{normalize}
\title{Normalize a list of data by dividing by it's
absolute max value.}
\usage{
normalize(values)
}
\arguments{
\item{values}{The list of values.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The normalized list.
}\if{html}{\out{</div>}}
}
\description{
Normalize a list of data by dividing by it's
absolute max value.
}
\examples{
normalize(c(0.9, 0.2, 0.3, 0.4, 0.1, 0.5))

}
