% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.r
\name{radar}
\alias{radar}
\title{Generates a radar graph.}
\usage{
radar(values, categories, bounds)
}
\arguments{
\item{values}{The y-coordinates.}

\item{categories}{The x-coordinates.}

\item{bounds}{A collection of the min and max bound.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The radar graph.
}\if{html}{\out{</div>}}
}
\description{
Generates a radar graph.
}
\examples{
radar(
  c(100, 0, 0, 0),
  c("A", "B", "C", "D"),
  c(0, 100)
)

}
