% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.r
\name{radar_anim}
\alias{radar_anim}
\title{Generates a radar graph animation between two values.}
\usage{
radar_anim(frames, start_values, end_values, categories, bounds)
}
\arguments{
\item{frames}{The number of frames to animate (60fps).}

\item{start_values}{The starting y-coordinates.}

\item{end_values}{The ending y-coordinates.}

\item{categories}{The x-coordinates.}

\item{bounds}{A collection of the min and max bound.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         The radar graph animation.
}\if{html}{\out{</div>}}
}
\description{
Generates a radar graph animation between two values.
}
\examples{
radar_anim(
  120,
  c(100, 0, 0, 0),
  c(25, 25, 25, 25),
  c("A", "B", "C", "D"),
  c(0, 100)
)

}
