% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcnoidanalysis-package.R
\docType{package}
\name{rcnoidanalysis-package}
\alias{rcnoidanalysis}
\alias{rcnoidanalysis-package}
\title{rcnoidanalysis: R CNoid Analysis}
\description{
Analyze and processess cnoid data.
}
\author{
\strong{Maintainer}: Tyler Nijmeh \email{tylernij@gmail.com}

}
\keyword{internal}
