% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{regress_cnoids_to_effects}
\alias{regress_cnoids_to_effects}
\title{Calculates the effect values for the given cnoid profile.}
\usage{
regress_cnoids_to_effects(d9 = 0, thcv = 0, cbd = 0, cbn = 0, cbg = 0, cbc = 0)
}
\arguments{
\item{d9}{D9 value.}

\item{thcv}{THCV value.}

\item{cbd}{CBD value.}

\item{cbn}{CBN value.}

\item{cbg}{CBG value.}

\item{cbc}{CBC value.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ The effect values.
}\if{html}{\out{</div>}}
}
\description{
Calculates the effect values for the given cnoid profile.
}
\examples{
regress_cnoids_to_effects(
  d9 = 100,
  thcv = 0,
  cbd = 0,
  cbn = 0,
  cbg = 0,
  cbc = 0
)

}
