% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_gen.r
\name{regress_effects_to_cnoids}
\alias{regress_effects_to_cnoids}
\title{Calculates the cnoid values for the given effect profile.}
\usage{
regress_effects_to_cnoids(
  aroused = 0,
  hungry = 0,
  energetic = 0,
  happy = 0,
  creative = 0,
  giggly = 0,
  uplifted = 0,
  focused = 0,
  talkative = 0,
  relaxed = 0,
  tingly = 0,
  euphoric = 0,
  sleepy = 0
)
}
\arguments{
\item{aroused}{Aroused value.}

\item{hungry}{Hungry value.}

\item{energetic}{Energetic value.}

\item{happy}{Happy value.}

\item{creative}{Creative value.}

\item{giggly}{Giggly value.}

\item{uplifted}{Uplifted value.}

\item{focused}{Focused value.}

\item{talkative}{Talkative value.}

\item{relaxed}{Relaxed value.}

\item{tingly}{Tingly value.}

\item{euphoric}{Euphoric value.}

\item{sleepy}{Sleepy value.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The cnoid values.
}\if{html}{\out{</div>}}
}
\description{
Calculates the cnoid values for the given effect profile.
}
\examples{
regress_effects_to_cnoids(
  aroused = 100,
  hungry = 0,
  energetic = 0,
  happy = 0,
  creative = 0,
  giggly = 0,
  uplifted = 0,
  focused = 0,
  talkative = 0,
  relaxed = 0,
  tingly = 0,
  euphoric = 0,
  sleepy = 0
)

}
