% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.r
\name{true_bounds_cnoids}
\alias{true_bounds_cnoids}
\title{Finds the true bounds for use with graphs
for the cnoids.}
\usage{
true_bounds_cnoids()
}
\value{
The min and max possible values that can
be generated from \code{generate_values()} for the
cnoids.
}
\description{
Finds the true bounds for use with graphs
for the cnoids.
}
