% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.r
\name{true_bounds_effects}
\alias{true_bounds_effects}
\title{Finds the true bounds for use with graphs
for the effects}
\usage{
true_bounds_effects()
}
\value{
The min and max possible values that can
be generated from \code{generate_values()} for the
effects.
}
\description{
Finds the true bounds for use with graphs
for the effects
}
