% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{anim_effects}
\alias{anim_effects}
\title{Generate a radar animation of the effects between two cnoid
profiles.}
\usage{
anim_effects(frames, start_values, end_values)
}
\arguments{
\item{frames}{The number of frames to animate (60fps).}

\item{start_values}{The starting cnoid values.}

\item{end_values}{The ending cnoid values.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         The graph animation.
}\if{html}{\out{</div>}}
}
\description{
The values should sum to 100.
}
\examples{
anim_effects(
  120,
  c(100, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
  c(50, 0, 0, 50, 0, 0, 0, 0, 0, 0, 0, 0, 0)
)

}
