% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{normalize_min_max}
\alias{normalize_min_max}
\title{Normalize a list of data by min-max method.}
\usage{
normalize_min_max(values, min, max)
}
\arguments{
\item{values}{The list of values.}

\item{max}{The maximum value.}

\item{main}{The minimum value.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The normalized list.
}\if{html}{\out{</div>}}
}
\description{
Normalize a list of data by min-max method.
}
\examples{
normalize_min_max(
  c(0.9, 0.2, 0.3, 0.4, 0.1, 0.5),
  0,
  1
)

}
